Refactor Vine related services
Move them under Vine module to keep the code nicely organised
