+++ spec/fixtures/vcr_cassettes/Vine_ApiService/_my_team/when_a_request_succeed/returns_the_response.yml
