+++ spec/fixtures/vcr_cassettes/Vine_ApiService/_voucher_validation/when_a_request_succeed/returns_the_response.yml
