+++ spec/requests/spree/admin/payments_spec.rb
@@ -76,16 +76,16 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
         create(:voucher_flat_rate, voucher_type: "VINE", code: 'some_code',
                                    enterprise: order.distributor, amount: 6)
       }
-      let(:vine_voucher_redeemer) { instance_double(VineVoucherRedeemerService) }
+      let(:vine_voucher_redeemer) { instance_double(Vine::VoucherRedeemerService) }
 
       before do
         add_voucher_to_order(vine_voucher, order)
 
-        allow(VineVoucherRedeemerService).to receive(:new).and_return(vine_voucher_redeemer)
+        allow(Vine::VoucherRedeemerService).to receive(:new).and_return(vine_voucher_redeemer)
       end
 
       it "completes the order and redirects to payment page" do
-        expect(vine_voucher_redeemer).to receive(:call).and_return(true)
+        expect(vine_voucher_redeemer).to receive(:redeem).and_return(true)
 
         post("/admin/orders/#{order.number}/payments.json", params:)
 
@@ -97,7 +97,7 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
 
       context "when redeeming the voucher fails" do
         it "redirect to payments page" do
-          allow(vine_voucher_redeemer).to receive(:call).and_return(false)
+          allow(vine_voucher_redeemer).to receive(:redeem).and_return(false)
           allow(vine_voucher_redeemer).to receive(:errors).and_return(
             { redeeming_failed: "Redeeming the voucher failed" }
           )
@@ -111,7 +111,7 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
 
       context "when an other error happens" do
         it "redirect to payments page" do
-          allow(vine_voucher_redeemer).to receive(:call).and_return(false)
+          allow(vine_voucher_redeemer).to receive(:redeem).and_return(false)
           allow(vine_voucher_redeemer).to receive(:errors).and_return(
             { vine_api: "There was an error communicating with the API" }
           )
@@ -267,16 +267,16 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
           create(:voucher_flat_rate, voucher_type: "VINE", code: 'some_code',
                                      enterprise: order.distributor, amount: 6)
         }
-        let(:vine_voucher_redeemer) { instance_double(VineVoucherRedeemerService) }
+        let(:vine_voucher_redeemer) { instance_double(Vine::VoucherRedeemerService) }
 
         before do
           add_voucher_to_order(vine_voucher, order)
 
-          allow(VineVoucherRedeemerService).to receive(:new).and_return(vine_voucher_redeemer)
+          allow(Vine::VoucherRedeemerService).to receive(:new).and_return(vine_voucher_redeemer)
         end
 
         it "completes the order and redirects to payment page" do
-          expect(vine_voucher_redeemer).to receive(:call).and_return(true)
+          expect(vine_voucher_redeemer).to receive(:redeem).and_return(true)
 
           put(
             "/admin/orders/#{order.number}/payments/#{order.payments.first.id}/" \
@@ -293,7 +293,7 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
 
         context "when redeeming the voucher fails" do
           it "redirect to payments page" do
-            allow(vine_voucher_redeemer).to receive(:call).and_return(false)
+            allow(vine_voucher_redeemer).to receive(:redeem).and_return(false)
             allow(vine_voucher_redeemer).to receive(:errors).and_return(
               { redeeming_failed: "Redeeming the voucher failed" }
             )
@@ -312,7 +312,7 @@ RSpec.describe Spree::Admin::PaymentsController, type: :request do
 
         context "when an other error happens" do
           it "redirect to payments page" do
-            allow(vine_voucher_redeemer).to receive(:call).and_return(false)
+            allow(vine_voucher_redeemer).to receive(:redeem).and_return(false)
             allow(vine_voucher_redeemer).to receive(:errors).and_return(
               { vine_api: "There was an error communicating with the API" }
             )
